fix types --WIP

+++ packages/trpc/server/routers/viewer/eventTypes/update.handler.ts
@@ -504,27 +504,34 @@ export const updateHandler = async ({ ctx, input }: UpdateOptions) => {
     }
   }
 
-  if (deletedWebhooks.length > 0) {
+  if (deletedWebhooks && deletedWebhooks.length > 0) {
     await ctx.prisma.$transaction(
-      deletedWebhooks.map((wh) =>
-        ctx.prisma.webhook.deleteMany({
-          where: {
-            id: wh.id,
-            eventTypeId: id,
-          },
-        })
-      )
+      deletedWebhooks
+        .filter((wh) => wh?.id)
+        .map((wh) =>
+          ctx.prisma.webhook.deleteMany({
+            where: {
+              id: wh.id,
+              eventTypeId: id,
+            },
+          })
+        )
     );
   }
 
-  if (webhooks) {
+  if (webhooks && webhooks.length > 0) {
     const webhooksForUpdates = webhooks.filter((wh) => wh.id);
-    const webhooksForInserts = webhooks
-      .filter((wh) => !wh.id) // New webhooks without id
-      .map((newWebhook) => ({
-        ...newWebhook,
-        id: v4(), // Generate a unique ID
-      }));
+    const webhooksForInserts = webhooks.reduce((acc, wh) => {
+      const subscriberUrl = wh.subscriberUrl || "";
+      if (!wh.id) {
+        acc.push({
+          ...wh,
+          subscriberUrl,
+          id: v4(), // Generate a unique ID
+        });
+      }
+      return acc;
+    }, [] as Array<{ id: string; subscriberUrl: string }>);
 
     await ctx.prisma.$transaction([
       ctx.prisma.webhook.createMany({
